@@ -14,8 +14,8 @@ description: A new Flutter application.
 version: 1.0.18+19
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: "2.2.0"
 
 dependencies:
   flutter:
@@ -25,42 +25,54 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
-  #ioc依赖注入
-  inject:
-    path: ./vendor/inject.dart/package/inject
   #lz_flutter 框架
   lz_flutter:
     git:
       url: http://gitlab.liangzhicn.com/lz-flutter/lz-flutter.git
-      ref: '1.2.0'
   #toast提示
-  fluttertoast: ^3.1.3
+  fluttertoast: ^8.0.7
   #网络请求
-  chopper: ^2.5.3
+  retrofit: ^2.0.0
+  dio: 4.0.0
   #时间格式
-  date_format: ^1.0.9
+  date_format: ^2.0.2
   #数据库
-  floor: ^0.18.0
+  floor: ^1.1.0
   synchronized: '>=1.5.1 <4.0.0'
   #缓存
-  shared_preferences: '>=0.4.1 <3.0.0'
+  shared_preferences: 2.0.6
   #路径
-  path_provider: '1.6.11'
-  json_annotation: ^3.0.0
+  path_provider: ^2.0.2
+  json_annotation: ^4.0.1
   #设备信息
-  device_info: 0.2.0
+  device_info: ^2.0.2
   #包信息
-  package_info: ^0.4.0+4
+  package_info: ^2.0.2
+  #指纹登录
+  local_auth: ^1.1.6
+  #第三方登录
+  firebase_auth: ^1.3.0
+  google_sign_in: ^5.0.4
+  sign_in_with_apple: ^3.0.0
+  #推送
+  flutter_local_notifications: ^6.0.0
+  flutter_apns: ^1.5.0
+  firebase_messaging: ^10.0.0
+  #依赖注入
+  injectable: ^1.4.1
+  get_it: ^7.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.8.1
-  chopper_generator: ^2.5.0
-  floor_generator: ^0.18.0
-  inject_generator:
-    path: ./vendor/inject.dart/package/inject_generator
-  json_serializable: ^3.2.0
+  build_runner: ^2.0.4
+  floor_generator: ^1.1.0
+  json_serializable: ^4.1.3
+  injectable_generator: ^1.4.1
+  retrofit_generator: ^2.0.0
+
+dependency_overrides:
+  firebase_messaging: ^10.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
