@@ -1,15 +1,14 @@
-import 'package:inject/inject.dart';
+import 'package:injectable/injectable.dart';
 
 import 'my_floor_database.dart';
 
 @singleton
-@provide
 class AppDataBase    {
 
-  MyFloorDatabase db;
+  late MyFloorDatabase db;
 
   Future<void> init() async =>
-    db = await $FloorMyFloorDatabase.databaseBuilder("app_database.db").build();
+    db = await $FloorMyFloorDatabase.databaseBuilder('app_database.db').build();
 
 
 }
