@@ -1,6 +1,6 @@
 import 'package:floor/floor.dart';
 import 'package:lz_flutter_app/network/domains/simple_http_request.dart';
-import 'package:lz_flutter_app/network/repositories/simple_http_request_local_repository.dart';
+import 'package:lz_flutter_app/network/local_persistence/simple_http_request_dao.dart';
 import 'package:sqflite/sqflite.dart' as sqflite;
 import 'package:path/path.dart';
 import 'dart:async';
@@ -9,5 +9,5 @@ part 'my_floor_database.g.dart';
 
 @Database(version: 1, entities: [SimpleHttpRequest])
 abstract class MyFloorDatabase extends FloorDatabase {
-  SimpleHttpRequestLocalRepository get httpRequestLocalRepository;
+  SimpleHttpRequestDao get simpleHttpRequestDao;
 }
