@@ -1,3 +1,3 @@
 //这个文件用于防止提交代码时忘记修改代理导致别的程序员运行出错
 //请在提交时 忽略这个文件
-const proxy = null;  //e.g. PROXY 192.168.31.41
+const proxy = null;  //e.g.  192.168.31.41:8888
